requirements.txt
streamlit
google-generativeai
pandas
Pillow
pdf2image

# !apt-get install -y poppler-utils
# !pip install pdf2image